pyproject.toml
[project]
name = "vghtpe-patient"
version = "0.1.0"
description = "VGHTPE patient data scraper with AI-powered clinical summary generation"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "pandas>=2.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "python-docx>=1.0",
    "rich>=13.0",
    "typer>=0.12",
]

[project.scripts]
vghtpe = "main:app"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends._legacy:_Backend"
